+++ ic-dbms-api/src/lib.rs
@@ -89,6 +89,22 @@
 //! - [`Page`](crate::memory::Page)
 //! - [`PageOffset`](crate::memory::PageOffset)
 //!
+//! ### Validate
+//!
+//! - [`Validate`](crate::prelude::Validate)
+//! - [`CamelCaseValidator`](crate::prelude::CamelCaseValidator)
+//! - [`CountryIso639Validator`](crate::prelude::CountryIso639Validator)
+//! - [`CountryIso3166Validator`](crate::prelude::CountryIso3166Validator)
+//! - [`EmailValidator`](crate::prelude::EmailValidator)
+//! - [`KebabCaseValidator`](crate::prelude::KebabCaseValidator)
+//! - [`MaxStrlenValidator`](crate::prelude::MaxStrlenValidator)
+//! - [`MimeTypeValidator`](crate::prelude::MimeTypeValidator)
+//! - [`MinStrlenValidator`](crate::prelude::MinStrlenValidator)
+//! - [`PhoneNumberValidator`](crate::prelude::PhoneNumberValidator)
+//! - [`RangeStrlenValidator`](crate::prelude::RangeStrlenValidator)
+//! - [`RgbColorValidator`](crate::prelude::RgbColorValidator)
+//! - [`SnakeCaseValidator`](crate::prelude::SnakeCaseValidator)
+//! - [`UrlValidator`](crate::prelude::UrlValidator)
 
 #![doc(html_playground_url = "https://play.rust-lang.org")]
 #![doc(
@@ -109,3 +125,4 @@ pub mod prelude;
 #[cfg(test)]
 mod tests;
 mod utils;
+mod validate;