+++ CHANGELOG.md
@@ -1,7 +1,15 @@
 # Changelog
 
 - [Changelog](#changelog)
+  - [0.2.0](#020)
   - [0.1.0](#010)
+
+## 0.2.0
+
+Released on 2025-12-21
+
+- [Field Validation](https://github.com/veeso/ic-dbms/pull/6): it is now possible to tag fields for validation. Validators can be specified in the schema and will be executed before inserting or updating records.
+  - The library comes with built-in validators for common use cases (e.g., email, URL, number range).
 
 ## 0.1.0